refactor(dev): split the dev-server registry into registry, workbenchLock and processLiveness
dev/registry/index.ts (449 lines) did three jobs. Split it into
siblings — registry.ts (the manifest store + watcher), workbenchLock.ts
(the workbench singleton lock) and processLiveness.ts (PID liveness +
start-time verification) — with index.ts re-exporting the existing
public API unchanged, so no call sites move.

The manifest schema previously extended workbenchLockSchema; its base
fields are now written out explicitly to keep the modules acyclic
(registry would otherwise import the lock's schema while the lock
imports the registry's dir + version). Parsed shape is identical.

Tests split along the same seams; the PID-reuse / Windows / startedAt
suites stay as integration tests through the public index API and live
in processLiveness.test.ts, matching the subject they guard.
