+++ packages/@sanity/cli/src/actions/dev/registry/__tests__/processLiveness.test.ts
@@ -1,3 +1,10 @@
+/**
+ * Tests for process-liveness behaviour (PID-reuse detection, OS-reported
+ * start times, the Windows PowerShell path) exercised through the registry
+ * and workbench-lock public API — these are integration tests by design, so
+ * they guard the plumbing between `processLiveness.ts` and its consumers,
+ * not just the helpers in isolation.
+ */
 import {existsSync, mkdirSync, readFileSync, writeFileSync} from 'node:fs'
 import {tmpdir} from 'node:os'
 import {join} from 'node:path'
@@ -11,7 +18,6 @@ import {
   getRegisteredServers,
   readWorkbenchLock,
   registerDevServer,
-  watchRegistry,
 } from '../index.js'
 
 const mockExecSync = vi.hoisted(() => vi.fn())
@@ -47,290 +53,6 @@ beforeEach(() => {
 afterEach(() => {
   vi.clearAllMocks()
   vi.unstubAllEnvs()
-})
-
-describe('registerDevServer', () => {
-  test('writes a manifest file and returns a cleanup function', () => {
-    const {release: cleanup} = registerDevServer({
-      host: 'localhost',
-      port: 3334,
-      type: 'studio',
-      workDir: '/tmp/project',
-    })
-
-    const filePath = join(registryDir(), `${process.pid}.json`)
-    expect(existsSync(filePath)).toBe(true)
-
-    const manifest = JSON.parse(readFileSync(filePath, 'utf8'))
-    expect(manifest.pid).toBe(process.pid)
-    expect(manifest.type).toBe('studio')
-    expect(manifest.port).toBe(3334)
-    expect(manifest.host).toBe('localhost')
-    expect(manifest.workDir).toBe('/tmp/project')
-    expect(manifest.startedAt).toBeDefined()
-    expect(manifest.version).toBe(1)
-
-    cleanup()
-    expect(existsSync(filePath)).toBe(false)
-  })
-
-  test('persists id in the manifest when provided', () => {
-    const {release: cleanup} = registerDevServer({
-      host: 'localhost',
-      id: 'app-abc',
-      port: 3334,
-      type: 'coreApp',
-      workDir: '/tmp/project',
-    })
-
-    const manifest = JSON.parse(readFileSync(join(registryDir(), `${process.pid}.json`), 'utf8'))
-    expect(manifest.id).toBe('app-abc')
-
-    cleanup()
-  })
-
-  test('persists projectId in the manifest when provided', () => {
-    const {release: cleanup} = registerDevServer({
-      host: 'localhost',
-      port: 3334,
-      projectId: 'x1g7jygt',
-      type: 'studio',
-      workDir: '/tmp/project',
-    })
-
-    const manifest = JSON.parse(readFileSync(join(registryDir(), `${process.pid}.json`), 'utf8'))
-    expect(manifest.projectId).toBe('x1g7jygt')
-
-    const servers = getRegisteredServers()
-    expect(servers).toHaveLength(1)
-    expect(servers[0].projectId).toBe('x1g7jygt')
-
-    cleanup()
-  })
-
-  test('omits optional metadata when not provided and retains manifest through getRegisteredServers', () => {
-    const {release: cleanup} = registerDevServer({
-      host: 'localhost',
-      port: 3334,
-      type: 'studio',
-      workDir: '/tmp/project',
-    })
-
-    const manifest = JSON.parse(readFileSync(join(registryDir(), `${process.pid}.json`), 'utf8'))
-    expect(manifest.id).toBeUndefined()
-    expect(manifest.manifest).toBeUndefined()
-
-    const servers = getRegisteredServers()
-    expect(servers).toHaveLength(1)
-    expect(servers[0].id).toBeUndefined()
-    expect(servers[0].manifest).toBeUndefined()
-
-    cleanup()
-  })
-
-  test('update inlines the extracted manifest into the registry entry', () => {
-    const {release: cleanup, update} = registerDevServer({
-      host: 'localhost',
-      port: 3334,
-      type: 'studio',
-      workDir: '/tmp/project',
-    })
-
-    const inlined = {createdAt: '2026-01-01T00:00:00.000Z', version: 3, workspaces: []}
-    update({manifest: inlined, manifestUpdatedAt: '2026-01-01T00:00:00.000Z'})
-
-    const servers = getRegisteredServers()
-    expect(servers[0].manifest).toEqual(inlined)
-    expect(servers[0].manifestUpdatedAt).toBe('2026-01-01T00:00:00.000Z')
-
-    cleanup()
-  })
-
-  test('cleanup does not throw if file already removed', () => {
-    const {release: cleanup} = registerDevServer({
-      host: 'localhost',
-      port: 3333,
-      type: 'studio',
-      workDir: '/tmp/project',
-    })
-
-    cleanup()
-    expect(() => cleanup()).not.toThrow()
-  })
-
-  test('update after release is a no-op — late background extractions do not re-create the file', () => {
-    const {release: cleanup, update} = registerDevServer({
-      host: 'localhost',
-      port: 3334,
-      type: 'studio',
-      workDir: '/tmp/project',
-    })
-
-    const filePath = join(registryDir(), `${process.pid}.json`)
-    cleanup()
-    expect(existsSync(filePath)).toBe(false)
-
-    // Simulate a background extraction completing after release.
-    update({
-      manifest: {createdAt: '2026-01-01T00:00:00.000Z', version: 3, workspaces: []},
-      manifestUpdatedAt: '2026-01-01T00:00:00.000Z',
-    })
-
-    expect(existsSync(filePath)).toBe(false)
-  })
-})
-
-describe('acquireWorkbenchLock', () => {
-  test('returns lock object when lock is available', () => {
-    const lock = acquireWorkbenchLock({host: 'localhost', port: 3333})
-    expect(lock).toBeDefined()
-    expect(lock!.release).toBeTypeOf('function')
-    expect(lock!.updatePort).toBeTypeOf('function')
-    lock!.release()
-  })
-
-  test('returns undefined when lock is already held by a live process', () => {
-    const lock = acquireWorkbenchLock({host: 'localhost', port: 3333})
-    expect(lock).toBeDefined()
-
-    const second = acquireWorkbenchLock({host: 'localhost', port: 3333})
-    expect(second).toBeUndefined()
-
-    lock!.release()
-  })
-
-  test('release removes the lock file', () => {
-    const lock = acquireWorkbenchLock({host: 'localhost', port: 3333})
-    lock!.release()
-
-    const second = acquireWorkbenchLock({host: 'localhost', port: 3333})
-    expect(second).toBeDefined()
-    second!.release()
-  })
-
-  test('stores host, port, startedAt and version in the lock file', () => {
-    const lock = acquireWorkbenchLock({host: '0.0.0.0', port: 4000})
-
-    const lockPath = join(registryDir(), 'workbench.lock')
-    const data = JSON.parse(readFileSync(lockPath, 'utf8'))
-    expect(data.host).toBe('0.0.0.0')
-    expect(data.port).toBe(4000)
-    expect(data.pid).toBe(process.pid)
-    expect(data.startedAt).toBeDefined()
-    expect(data.version).toBe(1)
-
-    lock!.release()
-  })
-
-  test('updatePort writes the new port to the lock file', () => {
-    const lock = acquireWorkbenchLock({host: 'localhost', port: 3333})
-    lock!.updatePort(3334)
-
-    const lockPath = join(registryDir(), 'workbench.lock')
-    const data = JSON.parse(readFileSync(lockPath, 'utf8'))
-    expect(data.port).toBe(3334)
-
-    lock!.release()
-  })
-
-  test('reclaims stale lock from a dead process', () => {
-    const dir = registryDir()
-    mkdirSync(dir, {recursive: true})
-
-    writeFileSync(
-      join(dir, 'workbench.lock'),
-      JSON.stringify({
-        host: 'localhost',
-        pid: 99_999_999,
-        port: 3333,
-        startedAt: new Date().toISOString(),
-        version: 1,
-      }),
-      {flag: 'wx'},
-    )
-
-    const lock = acquireWorkbenchLock({host: 'localhost', port: 3333})
-    expect(lock).toBeDefined()
-    lock!.release()
-  })
-
-  test('returns undefined after exhausting retries', () => {
-    const dir = registryDir()
-    mkdirSync(dir, {recursive: true})
-
-    // Write a lock that will fail schema validation (missing required fields),
-    // causing readWorkbenchLock to return undefined without pruning the file.
-    // With retries=0, the function should not recurse.
-    writeFileSync(join(dir, 'workbench.lock'), JSON.stringify({host: 'localhost', pid: 1, port: 1}))
-
-    const lock = acquireWorkbenchLock({host: 'localhost', port: 3333}, 0)
-    expect(lock).toBeUndefined()
-  })
-})
-
-describe('readWorkbenchLock', () => {
-  test('returns undefined when no lock file exists', () => {
-    expect(readWorkbenchLock()).toBeUndefined()
-  })
-
-  test('returns lock data when lock is held by a live process', () => {
-    const lock = acquireWorkbenchLock({host: '0.0.0.0', port: 4000})
-
-    const data = readWorkbenchLock()
-    expect(data).toBeDefined()
-    expect(data!.host).toBe('0.0.0.0')
-    expect(data!.port).toBe(4000)
-    expect(data!.pid).toBe(process.pid)
-
-    lock!.release()
-  })
-
-  test('prunes stale lock and returns undefined', () => {
-    const dir = registryDir()
-    mkdirSync(dir, {recursive: true})
-
-    const lockPath = join(dir, 'workbench.lock')
-    writeFileSync(
-      lockPath,
-      JSON.stringify({
-        host: 'localhost',
-        pid: 99_999_999,
-        port: 3333,
-        startedAt: new Date().toISOString(),
-        version: 1,
-      }),
-    )
-
-    expect(readWorkbenchLock()).toBeUndefined()
-    expect(existsSync(lockPath)).toBe(false)
-  })
-
-  // Regression: a zero-byte lock (e.g. left behind by a crashed writer or a
-  // killed `sanity dev` mid-write) used to early-return undefined without
-  // pruning, so the next acquire attempt hit EEXIST forever — `sanity dev`
-  // logged "Workbench dev server started at …" while no Vite was actually
-  // listening.
-  test('prunes zero-byte lock and returns undefined', () => {
-    const dir = registryDir()
-    mkdirSync(dir, {recursive: true})
-
-    const lockPath = join(dir, 'workbench.lock')
-    writeFileSync(lockPath, '')
-
-    expect(readWorkbenchLock()).toBeUndefined()
-    expect(existsSync(lockPath)).toBe(false)
-  })
-
-  test('prunes unparsable-JSON lock and returns undefined', () => {
-    const dir = registryDir()
-    mkdirSync(dir, {recursive: true})
-
-    const lockPath = join(dir, 'workbench.lock')
-    writeFileSync(lockPath, 'not json {{{')
-
-    expect(readWorkbenchLock()).toBeUndefined()
-    expect(existsSync(lockPath)).toBe(false)
-  })
 })
 
 describe('PID-reuse detection', () => {
@@ -650,56 +372,5 @@ describe('startedAt uses OS-reported process start time', () => {
     expect(storedMs).toBeLessThanOrEqual(after)
 
     cleanup()
-  })
-})
-
-describe('watchRegistry', () => {
-  test('invokes callback when a manifest file is added', async () => {
-    const callback = vi.fn()
-    const watcher = watchRegistry(callback)
-
-    const dir = registryDir()
-    const manifest: DevServerManifest = {
-      host: 'localhost',
-      pid: process.pid,
-      port: 5555,
-      startedAt: new Date().toISOString(),
-      type: 'studio',
-      version: 1,
-      workDir: '/tmp/watch-test',
-    }
-    writeFileSync(join(dir, `${process.pid}.json`), JSON.stringify(manifest))
-
-    await new Promise((resolve) => setTimeout(resolve, 500))
-
-    expect(callback).toHaveBeenCalled()
-    const servers = callback.mock.calls.at(-1)![0]
-    expect(servers.some((s: DevServerManifest) => s.port === 5555)).toBe(true)
-
-    watcher.close()
-  })
-
-  test('close stops notifications', async () => {
-    const callback = vi.fn()
-    const watcher = watchRegistry(callback)
-    watcher.close()
-
-    const dir = registryDir()
-    writeFileSync(
-      join(dir, 'after-close.json'),
-      JSON.stringify({
-        host: 'localhost',
-        pid: process.pid,
-        port: 6666,
-        startedAt: new Date().toISOString(),
-        type: 'studio',
-        version: 1,
-        workDir: '/tmp/closed',
-      }),
-    )
-
-    await new Promise((resolve) => setTimeout(resolve, 500))
-
-    expect(callback).not.toHaveBeenCalled()
   })
 })
