+++ packages/@sanity/cli/src/actions/dev/registry/index.ts
@@ -1,449 +1,15 @@
-import {execSync} from 'node:child_process'
-import {
-  existsSync,
-  mkdirSync,
-  readdirSync,
-  readFileSync,
-  unlinkSync,
-  watch,
-  writeFileSync,
-} from 'node:fs'
-import {join} from 'node:path'
-
-import {getSanityDataDir} from '@sanity/cli-core'
-import {z} from 'zod/mini'
-
-import {coreAppManifestSchema, studioManifestSchema} from '../../manifest/types.js'
-import {devDebug} from '../devDebug.js'
-import {canonicalizeWatchDir} from '../shared/canonicalizeWatchDir.js'
-
-/** Bump when the manifest/lock shape changes in a breaking way. */
-const REGISTRY_VERSION = 1
-
-const workbenchLockSchema = z.object({
-  host: z.string(),
-  pid: z.number(),
-  port: z.number(),
-  startedAt: z.string(),
-  version: z.literal(REGISTRY_VERSION),
-})
-
-const devServerManifestSchema = z.extend(workbenchLockSchema, {
-  id: z.optional(z.string()),
-  /**
-   * Interfaces the app exposes, mapped from the declared `views` (dock panels,
-   * `interface_type: "panel"`) and `services` (background workers,
-   * `interface_type: "worker"`). A service is just an interface, so both live
-   * in this one list. Carried separately from the manifest — interfaces live in
-   * the application service, not the manifest — so the workbench can render
-   * local panels and run local workers without a deploy. `entry_point` is the
-   * declared `src`. Lenient by design; the workbench is the authority on the
-   * interface shape.
-   */
-  interfaces: z.optional(
-    z.array(z.object({entry_point: z.string(), interface_type: z.string(), name: z.string()})),
-  ),
-  /** Inlined manifest — either a {@link StudioManifest} or {@link CoreAppManifest}. */
-  manifest: z.optional(z.union([studioManifestSchema, coreAppManifestSchema])),
-  /**
-   * ISO timestamp of the most recent successful manifest extraction. Bumped
-   * on every regeneration so re-writing this registry entry triggers the
-   * workbench `watchRegistry` watcher and forces a rebroadcast to clients.
-   */
-  manifestUpdatedAt: z.optional(z.string()),
-  projectId: z.optional(z.string()),
-  type: z.enum(['coreApp', 'studio']),
-  workDir: z.string(),
-})
 /**
- * A manifest describing a running dev server process (studio or app).
- * Stored as `~/.sanity/dev-servers/<pid>.json`.
+ * Registry of running dev server processes plus the workbench singleton lock.
  *
- * Workbench state is tracked separately via the lock file — see
- * {@link acquireWorkbenchLock} and {@link readWorkbenchLock}.
+ * - `registry.ts` — the store: one `<pid>.json` manifest per live dev server
+ * - `workbenchLock.ts` — the `workbench.lock` file guarding "one workbench per machine"
+ * - `processLiveness.ts` — PID liveness + start-time verification (PID-reuse detection)
  */
-export type DevServerManifest = z.infer<typeof devServerManifestSchema>
-
-/**
- * Returns the path to the dev server registry directory.
- * Uses the shared Sanity config directory to stay consistent with other CLI paths.
- */
-function getRegistryDir(): string {
-  return join(getSanityDataDir(), 'dev-servers')
-}
-
-/**
- * Check whether a process is still alive.
- * Sends signal 0 which doesn't kill anything — just checks existence.
- */
-function isProcessAlive(pid: number): boolean {
-  try {
-    process.kill(pid, 0)
-    return true
-  } catch (err: unknown) {
-    // EPERM means the process exists but we lack permission to signal it
-    if (err && typeof err === 'object' && 'code' in err && err.code === 'EPERM') {
-      return true
-    }
-    return false
-  }
-}
-
-/** Tolerance in ms when comparing stored vs OS-reported process start times. */
-const START_TIME_TOLERANCE_MS = 2000
-
-/**
- * Retrieve the OS-reported start time for a process.
- *
- * - Unix (macOS / Linux): `ps -o lstart=`
- * - Windows: PowerShell `Get-CimInstance Win32_Process` (`CreationDate`).
- *   We prefer CIM over `Get-Process -Id <pid>.StartTime` because `StartTime`
- *   opens a process handle and can throw "Access is denied" when the target
- *   is owned by another user, while `Win32_Process.CreationDate` is readable
- *   for any process the user can enumerate.
- *
- * On Windows the result for {@link process.pid} is memoised because the
- * start time of the current process never changes and PowerShell cold-start
- * is expensive (1–3s). Without the cache, a single `sanity dev` startup
- * spawns PowerShell 4–5 times for our own PID. Other PIDs are not cached —
- * they may be reused over time.
- *
- * Returns `undefined` if the process doesn't exist or the command fails.
- */
-let ownWindowsStartTime: {cached: true; date: Date | undefined} | undefined
-
-function getProcessStartTime(pid: number): Date | undefined {
-  if (process.platform === 'win32') {
-    if (pid === process.pid && ownWindowsStartTime) return ownWindowsStartTime.date
-    const result = readWindowsStartTime(pid)
-    if (pid === process.pid) ownWindowsStartTime = {cached: true, date: result}
-    return result
-  }
-  return readUnixStartTime(pid)
-}
-
-/** Test-only: clear the cached own-process start time. */
-export function __resetStartTimeCacheForTesting(): void {
-  ownWindowsStartTime = undefined
-}
-
-function readUnixStartTime(pid: number): Date | undefined {
-  try {
-    const output = execSync(`ps -o lstart= -p ${pid}`, {
-      encoding: 'utf8',
-      stdio: ['ignore', 'pipe', 'pipe'],
-      timeout: 1000,
-    }).trim()
-    if (!output) return undefined
-    const date = new Date(output)
-    return Number.isNaN(date.getTime()) ? undefined : date
-  } catch {
-    return undefined
-  }
-}
-
-function readWindowsStartTime(pid: number): Date | undefined {
-  try {
-    // `-NoProfile -NonInteractive` keeps PowerShell start-up minimal.
-    // `Get-CimInstance Win32_Process` exposes `CreationDate` for any process
-    // the user can enumerate — unlike `Get-Process.StartTime`, which opens
-    // a process handle and can fail with "Access is denied".
-    const output = execSync(
-      `powershell.exe -NoProfile -NonInteractive -Command "(Get-CimInstance Win32_Process -Filter 'ProcessId=${pid}').CreationDate.ToString('o')"`,
-      {encoding: 'utf8', stdio: ['ignore', 'pipe', 'pipe'], timeout: 5000},
-    ).trim()
-    if (!output) return undefined
-    const date = new Date(output)
-    return Number.isNaN(date.getTime()) ? undefined : date
-  } catch {
-    return undefined
-  }
-}
-
-/**
- * Check whether a process is alive **and** is the same process that wrote
- * the manifest/lock (not a PID that was reused by the OS after a crash).
- *
- * Compares the stored `startedAt` timestamp against the OS-reported process
- * start time. Falls back to a plain alive-check when the start time cannot
- * be retrieved (permissions, missing tools, etc.) — in that mode PID-reuse
- * goes undetected, but in practice {@link getProcessStartTime} succeeds on
- * all supported platforms.
- */
-function isOurProcess(pid: number, startedAt: string): boolean {
-  if (!isProcessAlive(pid)) return false
-
-  const osStart = getProcessStartTime(pid)
-  if (!osStart) return true // can't verify — assume alive is good enough
-
-  const storedStart = new Date(startedAt)
-  if (Number.isNaN(storedStart.getTime())) return true // bad stored value — fall back
-
-  return Math.abs(osStart.getTime() - storedStart.getTime()) <= START_TIME_TOLERANCE_MS
-}
-
-interface DevServerRegistration {
-  /** Remove the registry entry. */
-  release: () => void
-  /**
-   * Rewrite the registry entry with partial updates merged in. Also bumps the
-   * file's mtime, which fires `watchRegistry` in any workbench process and
-   * triggers a rebroadcast to connected clients.
-   */
-  update: (patch: Partial<Omit<DevServerManifest, 'pid' | 'startedAt' | 'version'>>) => void
-}
-
-/**
- * Write a manifest file for the current process and return a handle with a
- * `release` function that removes it plus an `update` function for patching
- * fields post-registration. Uses synchronous I/O so the file exists before
- * any signal handler could fire.
- */
-export function registerDevServer(
-  manifest: Omit<DevServerManifest, 'pid' | 'startedAt' | 'version'>,
-): DevServerRegistration {
-  const registryDir = getRegistryDir()
-  mkdirSync(registryDir, {recursive: true})
-
-  // Use the OS-reported process start time (falling back to now) so that
-  // `isOurProcess` can verify the manifest against the same reference on
-  // re-read. Using `new Date()` would record the manifest-write time, which
-  // drifts from the OS-reported process start by however long it took the
-  // process to reach this point — frequently exceeding START_TIME_TOLERANCE_MS
-  // and causing the manifest to be pruned as "stale" immediately after it's
-  // written.
-  let current: DevServerManifest = {
-    ...manifest,
-    pid: process.pid,
-    startedAt: (getProcessStartTime(process.pid) ?? new Date()).toISOString(),
-    version: REGISTRY_VERSION,
-  }
-
-  const filePath = join(registryDir, `${process.pid}.json`)
-  writeFileSync(filePath, JSON.stringify(current, null, 2))
-
-  // Guard against late updates from background tasks (e.g. the initial
-  // manifest extraction) landing after `release()` has deleted the file —
-  // without this, the update would re-create the registry entry and leak.
-  let released = false
-
-  return {
-    release() {
-      released = true
-      try {
-        unlinkSync(filePath)
-      } catch {
-        // ENOENT is fine — already cleaned up
-      }
-    },
-    update(patch) {
-      if (released) return
-      current = {...current, ...patch}
-      writeFileSync(filePath, JSON.stringify(current, null, 2))
-    },
-  }
-}
-
-/**
- * Read all manifest files from the registry, prune stale entries (dead PIDs),
- * and return the live ones.
- */
-export function getRegisteredServers(): DevServerManifest[] {
-  const registryDir = getRegistryDir()
-
-  if (!existsSync(registryDir)) {
-    return []
-  }
-
-  const files = readdirSync(registryDir).filter((f) => f.endsWith('.json'))
-  const servers: DevServerManifest[] = []
-
-  for (const file of files) {
-    const filePath = join(registryDir, file)
-    let raw: unknown
-    try {
-      raw = JSON.parse(readFileSync(filePath, 'utf8'))
-    } catch {
-      continue
-    }
-
-    const {data, success} = devServerManifestSchema.safeParse(raw)
-    if (!success) continue
-
-    if (isOurProcess(data.pid, data.startedAt)) {
-      servers.push(data)
-    } else {
-      // Prune stale manifest
-      try {
-        unlinkSync(filePath)
-      } catch {
-        // Ignore — another process may have already cleaned it up
-      }
-    }
-  }
-
-  return servers
-}
-
-/**
- * Read the workbench lock file and return its contents if the holding
- * process is still alive. Prunes stale locks from crashed processes.
- */
-export function readWorkbenchLock(): z.infer<typeof workbenchLockSchema> | undefined {
-  const lockPath = join(getRegistryDir(), 'workbench.lock')
-
-  let contents: string
-  try {
-    contents = readFileSync(lockPath, 'utf8')
-  } catch {
-    // File doesn't exist — nothing to prune, nothing to return
-    return undefined
-  }
-
-  // Past this point the file exists. Anything that isn't a live, valid lock
-  // (unparsable JSON, schema mismatch, dead/reused PID) is stale and must be
-  // pruned — otherwise the next `acquireWorkbenchLock` call is blocked by
-  // EEXIST forever and `sanity dev` silently no-ops the workbench server.
-  const data = parseLockContents(contents)
-  devDebug('Read workbench lock: %o', data)
-  if (data && isOurProcess(data.pid, data.startedAt)) {
-    devDebug('Workbench process is alive at pid %d on port %d', data.pid, data.port)
-    return data
-  }
-
-  pruneWorkbenchLock(lockPath)
-  return undefined
-}
-
-function parseLockContents(contents: string): z.infer<typeof workbenchLockSchema> | undefined {
-  try {
-    const {data, success} = workbenchLockSchema.safeParse(JSON.parse(contents))
-    return success ? data : undefined
-  } catch {
-    return undefined
-  }
-}
-
-function pruneWorkbenchLock(lockPath: string): void {
-  try {
-    devDebug('Removing stale workbench lock')
-    unlinkSync(lockPath)
-    devDebug('Stale workbench lock removed')
-  } catch {
-    // Another process may have already cleaned it up
-  }
-}
-
-interface WorkbenchLock {
-  /** Release the lock file. */
-  release: () => void
-  /** Update the lock with the actual port after the server starts listening. */
-  updatePort: (port: number) => void
-}
-
-/**
- * Attempt to acquire an exclusive lock for the workbench process.
- * Uses `O_EXCL` (the `wx` flag) which is atomic at the OS level — only one
- * process can create the file.
- *
- * The lock stores `{pid, host, port}` so other processes can find the
- * running workbench. Call `updatePort` after the Vite server starts to
- * write the actual port (Vite may pick a different one).
- *
- * @returns A {@link WorkbenchLock} if acquired, or `undefined` if another
- *          live process already holds it.
- */
-export function acquireWorkbenchLock(
-  info: {host: string; port: number},
-  retries = 1,
-): WorkbenchLock | undefined {
-  const registryDir = getRegistryDir()
-  mkdirSync(registryDir, {recursive: true})
-
-  const lockPath = join(registryDir, 'workbench.lock')
-  // Use OS-reported process start time for consistent comparison in
-  // `isOurProcess`. See the identical note in `registerDevServer`.
-  const startedAt = (getProcessStartTime(process.pid) ?? new Date()).toISOString()
-  const lockData = {
-    host: info.host,
-    pid: process.pid,
-    port: info.port,
-    startedAt,
-    version: REGISTRY_VERSION,
-  }
-
-  devDebug('Acquiring workbench lock at %s', lockPath)
-
-  try {
-    writeFileSync(lockPath, JSON.stringify(lockData), {flag: 'wx'})
-    devDebug('Workbench lock acquired')
-    return {
-      release() {
-        try {
-          unlinkSync(lockPath)
-        } catch {
-          // Already cleaned up
-        }
-      },
-      updatePort(port: number) {
-        writeFileSync(lockPath, JSON.stringify({...lockData, port}))
-      },
-    }
-  } catch (err: unknown) {
-    devDebug(
-      'Failed to acquire workbench lock: %s',
-      err instanceof Error ? err.message : String(err),
-    )
-    if (!isNodeError(err) || err.code !== 'EEXIST') return undefined
-
-    // Lock exists — check if the holder is still alive
-    const existing = readWorkbenchLock()
-    if (existing) return undefined
-
-    // Stale lock was pruned by readWorkbenchLock — retry (with guard against infinite recursion)
-    if (retries <= 0) return undefined
-    return acquireWorkbenchLock(info, retries - 1)
-  }
-}
-
-function isNodeError(err: unknown): err is NodeJS.ErrnoException {
-  return err instanceof Error && 'code' in err
-}
-
-interface RegistryWatcher {
-  close(): void
-}
-
-/**
- * Watch the registry directory for changes and invoke the callback with the
- * current list of live servers whenever a change is detected.
- *
- * Uses `fs.watch` with a debounce to coalesce rapid file changes (e.g. a
- * server starting and writing its manifest triggers multiple FS events).
- */
-export function watchRegistry(callback: (servers: DevServerManifest[]) => void): RegistryWatcher {
-  const registryDir = getRegistryDir()
-  mkdirSync(registryDir, {recursive: true})
-
-  // Canonicalize to the real long path so `fs.watch` doesn't abort on Windows
-  // short-path dirs. See `canonicalizeWatchDir`.
-  const watchDir = canonicalizeWatchDir(registryDir)
-
-  let debounceTimer: ReturnType<typeof setTimeout> | undefined
-
-  const notify = () => {
-    clearTimeout(debounceTimer)
-    debounceTimer = setTimeout(() => {
-      callback(getRegisteredServers())
-    }, 50)
-  }
-
-  const watcher = watch(watchDir, notify)
-
-  return {
-    close() {
-      clearTimeout(debounceTimer)
-      watcher.close()
-    },
-  }
-}
+export {__resetStartTimeCacheForTesting} from './processLiveness.js'
+export {
+  type DevServerManifest,
+  getRegisteredServers,
+  registerDevServer,
+  watchRegistry,
+} from './registry.js'
+export {acquireWorkbenchLock, readWorkbenchLock} from './workbenchLock.js'
